Order + factura

+++ Models/Order.cs
@@ -5,9 +5,9 @@
         public int Id { get; set; }
         public DateTime OrderDate { get; set; }
         public DateTime DevolutionDate { get; set; }
-        public double SubTotal { get; set; }
-        public double IVA { get; set; }
-        public double Total { get; set; }
+        public decimal SubTotal { get; set; }
+        public decimal IVA { get; set; }
+        public decimal Total { get; set; }
         public string Status { get; set; }   // Esto es para lo de pendiente, enviado, entregado, cancelado
         public int CustomerId { get; set; }
         public Customer Customer { get; set; }
