+++ Models/OrderDetail.cs
@@ -6,8 +6,8 @@
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public double Price { get; set; }
-        public double SubTotal { get; set; }
+        public decimal Price { get; set; }
+        public decimal SubTotal { get; set; }
         public Order Order { get; set; }
         public Product Product { get; set; }
     }
